Modification afin d'obtenir correctement la page dashboard.
