+++ controllers/Backend.php
@@ -7,6 +7,8 @@ class Backend
         $manager_comment = new CommentManager();
         $comments = $manager_comment->get_comments();
         
+        
+   
         //$manager_user = new Usermanager();
         //$comments = $manager_user->
         //$model_dashboard = new Dashboard();
